Trim root files and turn on the necessary branches only.
Trim the main root files and turn on the necessary branches used for neutrino-electron scattering analysis. This is especially useful for Montecarlo files as the file sizes are huge and hard to manage. Combining the root files without merging them cannot be possible at times if it gets bigger than 100 GB.